Change chatID type for TG bot from int to long

+++ src/main/java/ua/com/serverhelp/simplemetricstoragefile/alerter/sender/SimpleTelegramBot.java
@@ -27,7 +27,7 @@ public class SimpleTelegramBot implements AlertSender { //TODO add tests
     /**
      * work chat
      */
-    private int chatId;
+    private long chatId;
 
     @Override
     public void initialize(String jsonParams) throws JSONException {
